fix(nx-python): replace poetry update with poetry lock and poetry install when adding a new package
currently, when there are more than 2 levels of local projects dependencies,
when a dependency is added to the 2 level or deeper the nx python plugin uses
poetry update <lib> to update to keep all the poetry.lock in sync.

however, the poetry update only works well for the first level, for all the others
the poetry lock is not updated correctly.

the combination of poetry lock and only running poetry install when the root pyproject.toml
is not present, updates the poetry.lock for all the levels of dependencies.