feat(nx-python): preserve sorted-ness in root pyproject.toml
The `poetry-project` and `uv-project` generators make changes to the root `pyproject.toml` file, and they currently always add to the end of arrays/tables. This is not ideal for two reasons:

- if the array/table was previously sorted, it will (most likely) no longer be sorted
- always adding to the end essentially guarantees that any two changes will conflict with each other

If an array or table was previously sorted, insert the new element so as to keep the array/table sorted.